Patch for compatibility with SD Shield 3.0
Hello Marcello, thank you for the fantastic project!
I prepared the patch to make your code run on Arduino Due with this shield: http://wiki.iteadstudio.com/Stackable_SD_Card_shield_V3.0
I would be glad if you could test it with your shield wireless and see if it works. Without the patch SD.begin returns an error when communicating the shield and the software will not run.